[solver-ipopt] Changed file structure and renamed classes
