+++ CMakeLists.txt
@@ -128,6 +128,11 @@ IF(OPENMP_FOUND AND BUILD_WITH_MULTITHREADS)
   SET(OMP_NUM_THREADS ${BUILD_WITH_NTHREADS})
 ENDIF()
 
+# Add IpOpt
+IF(IPOPT_FOUND)
+  ADD_DEFINITIONS(-DCROCODDYL_WITH_IPOPT)
+ENDIF()
+
 SET(BOOST_REQUIERED_COMPONENTS filesystem serialization system)
 SET(BOOST_BUILD_COMPONENTS unit_test_framework)
 SET_BOOST_DEFAULT_OPTIONS()
@@ -166,12 +171,15 @@ IF(UNIX)
   SET_TARGET_PROPERTIES(${PROJECT_NAME} PROPERTIES LINKER_LANGUAGE CXX)
   TARGET_INCLUDE_DIRECTORIES(${PROJECT_NAME} PUBLIC $<INSTALL_INTERFACE:include>)
   TARGET_LINK_LIBRARIES(${PROJECT_NAME} pinocchio::pinocchio)
-  TARGET_LINK_LIBRARIES(${PROJECT_NAME} ipopt)
   TARGET_LINK_LIBRARIES(${PROJECT_NAME} ${Boost_FILESYSTEM_LIBRARY} ${Boost_SYSTEM_LIBRARY}
-      ${Boost_SERIALIZATION_LIBRARY})
-
+  ${Boost_SERIALIZATION_LIBRARY})
+  
   if(OPENMP_FOUND)
-    TARGET_LINK_LIBRARIES(${PROJECT_NAME} ${OpenMP_CXX_LIBRARIES})
+  TARGET_LINK_LIBRARIES(${PROJECT_NAME} ${OpenMP_CXX_LIBRARIES})
+  ENDIF()
+  
+  IF(IPOPT_FOUND)
+    TARGET_LINK_LIBRARIES(${PROJECT_NAME} ipopt)
   ENDIF()
 
   INSTALL(TARGETS ${PROJECT_NAME} EXPORT ${TARGETS_EXPORT_NAME} DESTINATION lib)
