+++ unittest/factory/solver.cpp
@@ -12,7 +12,7 @@
 #include "crocoddyl/core/solvers/fddp.hpp"
 #include "crocoddyl/core/solvers/box-ddp.hpp"
 #include "crocoddyl/core/solvers/box-fddp.hpp"
-#include "crocoddyl/core/solvers/ipopt/ipopt.hpp"
+#include "crocoddyl/core/solvers/ipopt.hpp"
 #include "crocoddyl/core/utils/exception.hpp"
 
 namespace crocoddyl {
@@ -37,8 +37,8 @@ std::ostream& operator<<(std::ostream& os, SolverTypes::Type type) {
     case SolverTypes::SolverBoxFDDP:
       os << "SolverBoxFDDP";
       break;
-    case SolverTypes::SolverIpOpt:
-      os << "SolverIpOpt";
+    case SolverTypes::SolverIpopt:
+      os << "SolverIpopt";
       break;
     case SolverTypes::NbSolverTypes:
       os << "NbSolverTypes";
@@ -87,8 +87,8 @@ boost::shared_ptr<crocoddyl::SolverAbstract> SolverFactory::create(SolverTypes::
     case SolverTypes::SolverBoxFDDP:
       solver = boost::make_shared<crocoddyl::SolverBoxFDDP>(problem);
       break;
-    case SolverTypes::SolverIpOpt:
-      solver = boost::make_shared<crocoddyl::SolverIpOpt>(problem);
+    case SolverTypes::SolverIpopt:
+      solver = boost::make_shared<crocoddyl::SolverIpopt>(problem);
       break;
     default:
       throw_pretty(__FILE__ ": Wrong SolverTypes::Type given");
